\name{chi.squared}
\alias{chi.squared}
\title{ Chi-squared filter }
\description{
  The algorithms find weights of discrete attributes basing on a chi-squared statistics.
}
\usage{
chi.squared(formula, data)
}
\arguments{
  \item{formula}{ a symbolic description of a model }
  \item{data}{ a symbolic description of a model }
}
\details{

}
\value{
a data.frame containing ranks for attributes in the first column and their names as row names
}
\author{ Piotr Romanski }
\examples{

}
