% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selector.info.gain.R
\name{entropy.based}
\alias{entropy.based}
\alias{gain.ratio}
\alias{gain.ratio,}
\alias{information.gain}
\alias{information.gain,}
\alias{symmetrical.uncertainty}
\title{Entropy-based filters}
\usage{
information.gain(formula, data, equalBins = 5)

gain.ratio(formula, data, equalBins = 5)

symmetrical.uncertainty(formula, data, equalBins = 5)
}
\arguments{
\item{formula}{a symbolic description of a model}

\item{data}{data to process - a \code{data.frame} or a \code{dgCMatrix} object}

\item{equalBins}{For how many equal bins to discretize an explanatory variable when it's not of type
\code{factor, character} or \code{logical}, when evaluating an entropy.}
}
\value{
A data.frame containing the worth of attributes in the first column and their names as row names
}
\description{
The algorithms find weights of discrete attributes basing on their correlation with continous class attribute.
}
\details{
\code{information.gain} is \deqn{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Class) + H(Attribute) - H(Class, Attribute)}

  \code{gain.ratio} is \deqn{\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute)}}{(H(Class) + H(Attribute) - H(Class, Attribute)) / H(Attribute)}

  \code{symmetrical.uncertainty} is \deqn{2\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute) + H(Class)}}{2 * (H(Class) + H(Attribute) - H(Class, Attribute)) / (H(Attribute) + H(Class))}
}
\examples{
# regular data.frame example
  data(iris)

  weights <- information.gain(Species~., iris)
  print(weights)
  subset <- cutoff.k(weights, 2)
  f <- as.simple.formula(subset, "Species")
  print(f)

  weights <- gain.ratio(Species~., iris)
  print(weights)
  subset <- cutoff.k(weights, 2)
  f <- as.simple.formula(subset, "Species")
  print(f)

  weights <- symmetrical.uncertainty(Species~., iris)
  print(weights)
  subset <- cutoff.biggest.diff(weights)
  f <- as.simple.formula(subset, "Species")
  print(f)

  # sparseMatrix example
}
\author{
Piotr Romanski,

Marcin Kosinski \link{sparseMatrix} format, \email{m.p.kosinski@gmail.com}
}

