\name{oneR}
\alias{oneR}
\title{ OneR algorithm }
\description{
  The algorithms find weights of discrete attributes basing on very simple association rules involving only one attribute in condition part.
}
\usage{
oneR(formula, data)
}
\arguments{
  \item{formula}{ a symbolic description of a model }
  \item{data}{ data to process }
}
\details{
  
}
\value{
a data.frame containing ranks for attributes in the first column and their names as row names
}
\author{ Piotr Romanski }
\examples{

}
